% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerFoldchange.R
\name{enhancerFoldchange}
\alias{enhancerFoldchange}
\title{Calculate the fold change of enhancers}
\usage{
enhancerFoldchange(
  e_df,
  se_df,
  permut_type = "normal",
  s1_pair = FALSE,
  s2_pair = FALSE,
  s1_r1_bam,
  s1_r2_bam,
  s2_r1_bam,
  s2_r2_bam
)
}
\arguments{
\item{e_df}{pooled macs2 output file *_peaks.narrowPeak of all samples.
Header needs to be "chr, start,end,name,score,strand,signalValue,pValue,qValue,peak"}

\item{se_df}{merged SE metadata from SEfilter}

\item{permut_type}{permutation type: stringent or normal (default=normal)}

\item{s1_pair}{if sample 1 is paired-end (default=FALSE)}

\item{s2_pair}{if sample 2 is paired-end (default=FALSE)}

\item{s1_r1_bam}{path of sample 1 replicate 1 bam file}

\item{s1_r2_bam}{path of sample 1 replicate 2 bam file}

\item{s2_r1_bam}{path of sample 2 replicate 1 bam file}

\item{s2_r2_bam}{path of sample 2 replicate 2 bam file}
}
\value{
A list of 2 datasets: enhancer foldchang results dataset and sizefactor used in DESeq2
}
\description{
A function to calculate the fold change of enhancers base on two conditions.
}
\details{
This function will extract the enhancers which in the range of SEs.
Then enhancers are merged if there gaps are within 500bps.
Feature count and DESeq2 are used to calculate the fold change of those enhancers
}
\examples{
foldchange_list <- enhancerFoldchange(pool_enhancer_df,se_meta,s1_r1_bam,s1_r2_bam,s2_r1_bam,s2_r2_bam)

}
