% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerFoldchange_bam.R
\name{enhancerFoldchange_bam}
\alias{enhancerFoldchange_bam}
\title{Calculate the fold change of enhancers}
\usage{
enhancerFoldchange_bam(
  e_df,
  se_df,
  c1_pair = FALSE,
  c2_pair = FALSE,
  condition_1,
  condition_2
)
}
\arguments{
\item{e_df}{pooled macs2 output file *_peaks.narrowPeak of all samples.
Header needs to be "chr, start,end,name,score,strand,signalValue,pValue,qValue,peak"}

\item{se_df}{merged SE metadata from SEfilter}

\item{c1_pair}{if replicates of condition 1 is paired-end, separated by ",".
Order need to match with parameter condition_1.
Example: c1_pair="FALSE(or F),TRUE(or T),..."}

\item{c2_pair}{if replicates of condition 1 is paired-end, separated by ",".
Order need to match with parameter condition_2.
Example: c2_pair="FALSE(or F),TRUE(or T),..."}

\item{condition_1}{path of bam files for condition 1 replicates, separated by ",".
Example: condition_1="path_to_bam/c1_r1.bam,path_to_bam/c1_r2.bam,...".}

\item{condition_2}{path of bam files for condition 1 replicates, separated by ",".
Example: condition_2="path_to_bam/c2_r1.bam,path_to_bam/c2_r2.bam,...".}
}
\value{
A list of 2 datasets: enhancer foldchang results dataset and sizefactor used in DESeq2
}
\description{
A function to calculate the fold change of enhancers base on two conditions.
}
\details{
This function will extract the enhancers which in the range of SEs.
Then enhancers are merged if there gaps are within 500bps.
Feature count and DESeq2 are used to calculate the fold change of those enhancers
}
\examples{
foldchange_list <- enhancerFoldchange_bam(pool_enhancer_df,se_meta,
                       condition_1="path_to_bam/c1_r1.bam,path_to_bam/c1_r2.bam,...",
                       condition_2="path_to_bam/c2_r1.bam,path_to_bam/c2_r2.bam,...",
                       c1_pair="F,T,...",
                       c2_pair="F,T,...")

}
