% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerFoldchange_count.R
\name{enhancerFoldchange_count}
\alias{enhancerFoldchange_count}
\title{Calculate the fold change of enhancers}
\usage{
enhancerFoldchange_count(e_df, se_df, raw_count, c1_n, c2_n)
}
\arguments{
\item{e_df}{pooled macs2 output file *_peaks.narrowPeak of all samples.}

\item{se_df}{merged SE metadata from SEfilter.}

\item{raw_count}{raw count matrix of pooled enhancers in each sample.
The order of sample columns need to be "condition1_rep1, condition1_rep2,...,condition2_rep1,condition2_rep2"}

\item{c1_n}{number of replicates (samples) in condition 1.}

\item{c2_n}{number of replicates (samples) in condition 2.}
}
\value{
A list of 2 datasets: enhancer foldchang results dataset and sizefactor used in DESeq2
}
\description{
A function to calculate the fold change of enhancers base on two conditions.
}
\details{
This function will extract the enhancers which in the range of SEs with count table of pooled enhancers.
Then enhancers are merged with no gaps.
DESeq2 is used to calculate the fold change of those enhancers
}
\examples{
foldchange_list <- enhancerFoldchange_count(e_df,se_df,raw_count,c1_n=2,c2_n=2)

}
