% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEcategory.R
\name{SEcategory}
\alias{SEcategory}
\title{Category and ranking of SE}
\usage{
SEcategory(se_seg_df, e_fit)
}
\arguments{
\item{se_seg_df}{SE segment percentage file form SEpattern}

\item{e_fit}{enhancer deseq out file from enhancerFoldchange}
}
\value{
A dataset of final SE category and ranking. Each column represents the following:

se_merge_name: name of merged SE,"chr_start_end".
total_width: width of merged SE (unit=k).
number_enhancer: number of CEs in each SE.
category: SE category identified by \emph{DASE}.
direction: enrichment direction of SEs (none:Other or non-differential category; +: enriched in sample 2; -: enriched in sample 1; l:  sample 1 shifted in 5' direction; r: sample 2 shifted in 5' direction).
non_mid_percent: percentage of non middle segments.
mean_FC: mean of log2 SE coverage fold change.
rank: SE category ranking based on non_mid_percent first and mean_FC for each SE category. (rank=1 means the most like to the SE category.)
}
\description{
A function to get the category of SEs based on SE segment percentage file
}
\details{
This function will create the category of SEs (similar, strengthen/weaken, shorten,
shifting, V-shape, other) based on the segment percentage. Within each category, SE are
firstly ranked by segment percentage, then by average fold change of enhancers within
}
\examples{
catRank_list <- SEcategory(segment_percentage_df,enhancerFoldchange_out)

}
