% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEcategory_demo_2.R
\name{SEcategory_demo_2}
\alias{SEcategory_demo_2}
\title{Category and ranking of SE}
\usage{
SEcategory_demo_2(se_seg_df, e_fit)
}
\arguments{
\item{se_seg_df}{SE segment percentage file form SEpattern}

\item{e_fit}{enhancer deseq out file from enhancerFoldchange}
}
\value{
A dataset of final SE category and ranking.
}
\description{
A function to get the category of SEs based on SE segment percentage file
}
\details{
This function will create the category of SEs (similar, strengthen/weaken, shorten,
shifting, V-shape, other) based on the segment percentage. Within each category, SE are
firstly ranked by segment percentage, then by average fold change of enhancers within
}
\examples{
catRank_list <- SEcategory(segment_percentage_df,enhancerFoldchange_out)

}
