% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEfilter.R
\name{SEfilter}
\alias{SEfilter}
\title{Filter super-enhancer}
\usage{
SEfilter(se_in, bl_file = FALSE, has_bl_file = FALSE)
}
\arguments{
\item{se_in}{pooled ROSE output file *_peaks_Gateway_SuperEnhancers.bed of all samples.
Header needs to be "CHROM,START,STOP,REGION_ID,Signal,STRAND"}

\item{bl_file}{super-enhancer blacklist bed file download from ENCODE (ENCFF356LFX) (default=FALSE)}

\item{has_bl_file}{if there is a blacklist file (default=FALSE)}
}
\value{
A list of 3 datasets: filtered SE datasets with merged SE names,
SE datasets identified in the blacklist, and merged SE metadata
}
\description{
A function to filter super-enhancers (SEs)
}
\details{
This function firstly remove SEs whose width are within lower and upper 1 percent.
Then filter out SEs in the range of super-enhancer blacklist.
Finally, remained SEs from different samples are merged to the longest representative SE.
}
\examples{
# run without blacklist file
se_list <- SEfilter(se_in)

# run with blacklist file
se_list <- SEfilter(se_in,bl_file,has_bl_file=TRUE)

}
