% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEfitspline_new.R
\name{SEfitspline_new}
\alias{SEfitspline_new}
\title{Fit enhancer fold change and normal premut}
\usage{
SEfitspline_new(e_deseq_out_df)
}
\arguments{
\item{e_deseq_out_df}{enhancer output file from enhancerFoldchange}

\item{times}{permutation times (default=10)}

\item{permut}{if you want permutation (default=TRUE)}
}
\value{
A list of 4 datasets: enhancer fitted dataset, permutation dataset, cutoff vector,
and a density plot of permutation and original fitted fold change.
}
\description{
A function uses bs-spline weighted by basemean to fit the enhancer fold change
within each SE.
}
\details{
This function will fit the enhancer fold change within each SE by using bs-spline
weighted by basemean. A normal permutation of counts of enhancers in SEs is used to decide the
cutoff of significant log2FC (for samples are not too similar). Cutoff was calculate by slope equal to range(y)/range(x).
The max value in range(-2,-1.5) and min value in range(1.5,2) are set to be lower and upper cutoff.
}
\examples{
# permutation 10 times
fit_list <- SEfitspline(enhancerFoldchange_out,se_meta)

# permutation 5 times
fit_list <- SEfitspline(enhancerFoldchange_out,se_meta,times=5)

# no permutation
fit_list <- SEfitspline(enhancerFoldchange_out,se_meta,permut=FALSE)
}
