% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEmain.R
\name{SEmain}
\alias{SEmain}
\title{Get SE category by one step}
\usage{
SEmain(
  se_in,
  e_df,
  bl_file = FALSE,
  has_bl_file = FALSE,
  permut = TRUE,
  times = 10,
  s1_r1_bam,
  s1_r2_bam,
  s2_r1_bam,
  s2_r2_bam
)
}
\arguments{
\item{se_in}{pooled ROSE output file *_peaks_Gateway_SuperEnhancers.bed of all samples.
Header needs to be "CHROM,START,STOP,REGION_ID,Signal,STRAND"}

\item{e_df}{pooled macs2 output file *_peaks.narrowPeak of all samples.
Header needs to be "chr, start,end,name,score,strand,signalValue,pValue,qValue,peak"}

\item{bl_file}{super-enhancer blacklist bed file download from ENCODE (ENCFF356LFX) (default=FALSE)}

\item{has_bl_file}{if there is a blacklist file (default=FALSE)}

\item{permut}{if you want permutation (default=TRUE)}

\item{times}{permutation times (default=10)}

\item{s1_r1_bam}{path of sample 1 replicate 1 bam file}

\item{s1_r2_bam}{path of sample 1 replicate 2 bam file}

\item{s2_r1_bam}{path of sample 2 replicate 1 bam file}

\item{s2_r2_bam}{path of sample 2 replicate 2 bam file}
}
\value{
A list of 6 datasets: permutation density plot, pattern plots of each SE,
enhancer fitted dataset, fold change cutoff vector,
SE segment percentage dataset, and SE category./ranking dataset
}
\description{
A function to get the category and ranking of SEs in one step
}
\details{
This function is the main function used to get the category and ranking of SEs.
The workflow of this function is "SEfilter -> enhancerFoldchange -> SEfitspline -> SEpattern -> SEcategory"
}
\examples{
# no blacklist with permutation 10 times
se_main_list <- SEmain(se_in=pooled_rose,e_df=pooled_enhancer,
s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,
s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path)

# with blacklist and permutation 10 times
se_main_list <- SEmain(se_in=pooled_rose,e_df=pooled_enhancer,bl_file= blacklist,
has_bl_file=TRUE, s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path)

# no blacklist nor permutation 10 times
se_main_list <- SEmain(se_in=pooled_rose,e_df=pooled_enhancer,permut=FALSE,
s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path)

}
