% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEpattern_demo.R
\name{SEpattern_demo}
\alias{SEpattern_demo}
\title{Pattern of SE}
\usage{
SEpattern_demo(se_fit_df, pattern_plot_df, cutoff_vector)
}
\arguments{
\item{se_fit_df}{fitted output file from SEfitspline}

\item{cutoff_vector}{cutoff vector including lower and upper cutoff from SEfitspline
(default is c(-1.5,1.5))}
}
\value{
A list of 2 datasets: plot list contain plots' object of each SE and
SE segment percentage dataset
}
\description{
A function to get the pattern of SEs based on enhancer fitted line
}
\details{
This function will generate plot patterns of each SE and based on the intersection of
cutoff and fitted line, patterns are seperated into several segments. The percentage of
each segments are calculated by basemean of enhancer/sum of basemean within each SE
}
\examples{
seg_list <- SEpattern(SEfitspline_out,SEfitspline_cutoff)

}
