% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEpermut.R
\name{SEpermut}
\alias{SEpermut}
\title{Significance estimation with permutation}
\usage{
SEpermut(spline_fit_out, sample_pool, permut = T, times = 10)
}
\arguments{
\item{spline_fit_out}{enhancer output file from SEfitspline}

\item{sample_pool}{enhancer counts shuffling pool
(header must be "e_merge_name","S1_r1","S1_r2","S2_r1","S2_r2", "se_merge_name"))}

\item{permut}{if you want permutation (default=TRUE)}

\item{times}{permutation times (default=10)}
}
\value{
se_permutation_df: permutation data frame
cutoff: final cutoff
cutoff_candidate: all cutoff candidates
density_plot: density plot of permuted and original log2 fold change
}
\description{
A function to estimate the significance cutoff with permutation
}
\details{
A permutation of counts of enhancers in SEs is used to decide the
cutoff of significant log2FC with shuffling pool. Cutoff was calculate by slope equal to range(y)/range(x).
The max value in range(-2,-1) and min value in range(1,2) are set to be lower and upper cutoff.
}
\examples{
# default permutation 10 times
permut_out <- SEpermut(fit_spline_out_df,sample_pool)

# permutation 5 times
permut_out <- SEpermut(fit_spline_out_df,sample_pool,times=5)

# no permutation
permut_out <- SEpermut(fit_spline_out_df,sample_pool,permut=F)
}
