% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASE.R
\name{DASE}
\alias{DASE}
\title{Main function to estimate differential SEs}
\usage{
DASE(
  se_in,
  e_in,
  bl_file,
  custom_range,
  enhancer_count_table,
  data_type = "bam",
  s1_pair = F,
  s2_pair = F,
  spline_fun = "bs",
  permut = T,
  times = 10,
  cutoff_v = c(-1, 1),
  s1_r1_bam,
  s1_r2_bam,
  s2_r1_bam,
  s2_r2_bam
)
}
\arguments{
\item{se_in}{pooled SE bed file of all samples.}

\item{e_in}{pooled enhancer bed file of all samples.}

\item{bl_file}{super-enhancer blacklist bed file download from ENCODE (ENCFF356LFX)}

\item{custom_range}{a vector of extra customized blacklist to ignore.
Format: c(chr:start-end, chr:start-end, ...).}

\item{enhancer_count_table}{raw count file of pooled enhancers}

\item{data_type}{quantitative file format "bam" or "bw" (default=bam)}

\item{s1_pair}{if sample 1 is paired-end (default=FALSE)}

\item{s2_pair}{if sample 2 is paired-end (default=FALSE)}

\item{spline_fun}{spline functions ("bs", "ns", and "smooth". default=bs)}

\item{permut}{if you want permutation (default=TRUE)}

\item{times}{permutation times (default=10)}

\item{cutoff_v}{fold change lower and upper cutoff vector.
if permutation is used, it will use the value calculated by permutation. (default=c(-1.5,1.5))}

\item{s1_r1_bam}{path of sample 1 replicate 1 bam/bw file}

\item{s1_r2_bam}{path of sample 1 replicate 2 bam/bw file}

\item{s2_r1_bam}{path of sample 2 replicate 1 bam/bw file}

\item{s2_r2_bam}{path of sample 2 replicate 2 bam/bw file}
}
\value{
The output of DASE is a list with multiple data types including:

lfc_shrink: a shrinking lfc object from \emph{DESeq2} for all enhancers. It can be used to creat MA plot
cutoff: significant threshold for fitted log2 fold changes.
density_plot: a density plot of permutation and original fitted log2 fold changes, if \emph{permut=T}.
boxplot: a boxplot of final SE categories
se_category: a data frame containing final SE categories
pattern_list: a list containing figures for each SE pattern
ce_fit: a data frame containing DESeq2 output and spline-fitted log2 fold change of all constitute enhancers

Each column in se_category represents:

se_merge_name: name of merged SE,"chr_start_end".
total_width: width of merged SE (unit=k).
number_enhancer: number of CEs in each SE.
category: SE category identified by \emph{DASE}.
direction: enrichment direction of SEs (none:Other or non-differential category; +: enriched in sample 2; -: enriched in sample 1; l:  sample 1 shifted in 5' direction; r: sample 2 shifted in 5' direction).
non_mid_percent: percentage of non middle segments.
mean_FC: mean of log2 SE coverage fold change.
rank: SE category ranking based on non_mid_percent first and mean_FC for each SE category. (rank=1 means the most like to the SE category.)

Each column in ce_fit represents:

e_merge_name: name of merged CEs.
chr: chromosome of CEs.
start: start posation of CEs.
end: end posation of CEs.
width: width of CEs.
S1_r1: coverage of sample 1 replicate 1.
S1_r2: coverage of sample 1 replicate 2.
S2_r1: coverage of sample 2 replicate 1.
S2_r2: coverage of sample 2 replicate 2.
S1_r1_norm: normalized coverage of sample 1 replicate 1.
S1_r2_norm: normalized coverage of sample 1 replicate 2.
S2_r1_norm: normalized coverage of sample 2 replicate 1.
S2_r2_norm: normalized coverage of sample 2 replicate 2.
baseMean: normalized coverage mean of all samples.
log2FoldChange: log2 fold change of normalized coverage.
lfcSE: log2 fold change standard error.
stat: statistic of log2 fold change.
pvalue: p-value of log2 fold change.
padj: adjusted p-value of log2 fold change.
baseMean_shrink: shrinkage of normalized coverage mean of all samples.
log2FoldChange_shrink: shrinkage of log2 fold change..
lfcSE_shrink: shrinkage of log2 fold change standard error.
pvalue_shrink: shrinkage of log2 fold change p-value
padj_shrink: shrinkage of log2 fold change adjusted p-value
se_merge_name: merged SE names
}
\description{
A function to estimate and rank  differential SEs in one step
}
\details{
This function is the main function used to estimate the differential category and ranking of SEs.
The workflow of this function is "SEfilter -> enhancerFoldchange -> SEfitspline -> SEpattern -> SEcategory"
}
\examples{
# no blacklist with permutation 10 times with b-spline function
se_main_list <- DASE(se_in=pooled_se,e_in=pooled_enhancer,
s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,
s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path)

# with blacklist and permutation 10 times with n-spline function
se_main_list <- DASE(se_in=pooled_se,e_in=pooled_enhancer,bl_file= blacklist,
s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path,
spline_fun="ns")

# no blacklist nor permutation 10 times with smooth spline function
se_main_list <- DASE(se_in=pooled_se,e_in=pooled_enhancer,permut=FALSE,
s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path,
spline_fun="smooth")

}
