% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASE.R
\name{DASE}
\alias{DASE}
\title{Main function to get SE category}
\usage{
DASE(
  se_in,
  e_in,
  bl_file,
  custom_range,
  enhancer_count_table,
  data_type = "bam",
  s1_pair = F,
  s2_pair = F,
  permut = T,
  times = 10,
  cutoff_v = c(-1, 1),
  s1_r1_bam,
  s1_r2_bam,
  s2_r1_bam,
  s2_r2_bam
)
}
\arguments{
\item{se_in}{pooled ROSE output file *_peaks_Gateway_SuperEnhancers.bed of all samples.}

\item{e_in}{pooled macs2 output file *_peaks.narrowPeak of all samples.}

\item{bl_file}{super-enhancer blacklist bed file download from ENCODE (ENCFF356LFX)}

\item{custom_range}{a vector of extra customized blacklist to ignore.
Format: c(chr:start-end, chr:start-end, ...).}

\item{enhancer_count_table}{raw count file of pooled enhancers}

\item{data_type}{quantitative file format "bam" or "bw" (default=bam)}

\item{s1_pair}{if sample 1 is paired-end (default=FALSE)}

\item{s2_pair}{if sample 2 is paired-end (default=FALSE)}

\item{permut}{if you want permutation (default=TRUE)}

\item{times}{permutation times (default=10)}

\item{cutoff_v}{fold change lower and upper cutoff vector.
if permutation is used, it will use the value calculated by permutation. (default=c(-1.5,1.5))}

\item{s1_r1_bam}{path of sample 1 replicate 1 bam/bw file}

\item{s1_r2_bam}{path of sample 1 replicate 2 bam/bw file}

\item{s2_r1_bam}{path of sample 2 replicate 1 bam/bw file}

\item{s2_r2_bam}{path of sample 2 replicate 2 bam/bw file}
}
\value{
A list of 7 datasets: lfc_shrink, permutation density plot, pattern plots of each SE,
enhancer fitted dataset, significant threshold cutoff, SE category and boxplot of SE category
}
\description{
A function to get the category and ranking of SEs in one step
}
\details{
This function is the main function used to get the category and ranking of SEs.
The workflow of this function is "SEfilter -> enhancerFoldchange -> SEfitspline -> SEpattern -> SEcategory"
}
\examples{
# no blacklist with permutation 10 times
se_main_list <- SEprofile(se_in=pooled_rose,e_in=pooled_enhancer,
s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,
s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path)

# with blacklist and permutation 10 times
se_main_list <- SEprofile(se_in=pooled_rose,e_in=pooled_enhancer,bl_file= blacklist,
s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path)

# no blacklist nor permutation 10 times
se_main_list <- SEprofile(se_in=pooled_rose,e_in=pooled_enhancer,permut=FALSE,
s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path)

}
