% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DASE.R
\name{DASE}
\alias{DASE}
\title{Main function to get SE category}
\usage{
DASE(
  se_in,
  e_in,
  bl_file,
  custom_range,
  enhancer_count_table,
  data_type = "bam",
  spline_fun = "bs",
  permut = T,
  times = 10,
  cutoff_v = c(-1, 1),
  condition_1,
  condition_2,
  c1_pair,
  c2_pair,
  c1_n,
  c2_n
)
}
\arguments{
\item{se_in}{pooled SE bed file of all samples.}

\item{e_in}{pooled enhancer bed file of all samples.}

\item{bl_file}{super-enhancer blacklist bed file download from ENCODE (ENCFF356LFX)}

\item{custom_range}{a vector of extra customized blacklist to ignore.
Format: c(chr:start-end, chr:start-end, ...).}

\item{enhancer_count_table}{raw count file of pooled enhancers}

\item{data_type}{quantitative file format "bam" or "bw" (default=bam)}

\item{spline_fun}{spline functions ("bs", "ns", and "smooth". default=bs)}

\item{permut}{if you want permutation (default=TRUE)}

\item{times}{permutation times (default=10)}

\item{cutoff_v}{fold change lower and upper cutoff vector.
if permutation is used, it will use the value calculated by permutation. (default=c(-1.5,1.5))}

\item{condition_1}{path of bam/bw files for condition 1 replicates, separated by ",".
Example: condition_1="path_to_bam/c1_r1.bam,path_to_bam/c1_r2.bam,...".}

\item{condition_2}{path of bam/bw files for condition 1 replicates, separated by ",".
Example: condition_2="path_to_bam/c2_r1.bam,path_to_bam/c2_r2.bam,...".}

\item{c1_pair}{if replicates of condition 1 is paired-end, separated by ",".
Order need to match with parameter condition_1.
Example: c1_pair="FALSE(or F),TRUE(or T),...". (only available for bam file)}

\item{c2_pair}{if replicates of condition 1 is paired-end, separated by ",".
Order need to match with parameter condition_2.
Example: c2_pair="FALSE(or F),TRUE(or T),...". (only available for bam file)}

\item{c1_n}{number of replicates (samples) in condition 1 (only available for raw count table).}

\item{c2_n}{number of replicates (samples) in condition 2 (only available for raw count table).}
}
\value{
The output of DASE is a list with multiple data types including:

lfc_shrink: a shrinking lfc object from \emph{DESeq2} for all enhancers. It can be used to creat MA plot
cutoff: significant threshold for fitted log2 fold changes.
density_plot: a density plot of permutation and original fitted log2 fold changes, if \emph{permut=T}.
boxplot: a boxplot of final SE categories
se_category: a data frame containing final SE categories
pattern_list: a list containing figures for each SE pattern
ce_fit: a data frame containing DESeq2 output and spline-fitted log2 fold change of all constitute enhancers

Each column in se_category represents:

se_merge_name: name of merged SE,"chr_start_end".
total_width: width of merged SE (unit=k).
number_enhancer: number of CEs in each SE.
category: SE category identified by \emph{DASE}.
direction: enrichment direction of SEs (none:Other or non-differential category; +: enriched in sample 2; -: enriched in sample 1; l: sample 1 shifted in 5' direction; r: sample 2 shifted in 5' direction).
non_mid_percent: total activity occupancy of the segments that go beyond the threshold cutoffs.
mean_FC: mean of log2 SE coverage fold change.
rank: SE category ranking based on non_mid_percent first and mean_FC for each SE category. (rank=1 means the most changed in the corresponding SE category.)

Each column in ce_fit represents:

e_merge_name: name of merged CEs.
chr: chromosome of CEs.
start: start posation of CEs.
end: end posation of CEs.
width: width of CEs.
C1_1: coverage of condition 1 replicate 1.
C1_2: coverage of condition 1 replicate 2.
C2_1: coverage of condition 2 replicate 1.
C2_2: coverage of condition 2 replicate 2.
C1_1_norm: normalized coverage of condition 1 replicate 1.
C1_2_norm: normalized coverage of condition 1 replicate 2.
C2_1_norm: normalized coverage of condition 2 replicate 1.
C2_2_norm: normalized coverage of condition 2 replicate 2.
baseMean: normalized coverage mean of all samples.
log2FoldChange: log2 fold change of normalized coverage.
lfcSE: log2 fold change standard error.
stat: statistic of log2 fold change.
pvalue: p-value of log2 fold change.
padj: adjusted p-value of log2 fold change.
baseMean_shrink: shrinkage of normalized coverage mean of all samples.
log2FoldChange_shrink: shrinkage of log2 fold change..
lfcSE_shrink: shrinkage of log2 fold change standard error.
pvalue_shrink: shrinkage of log2 fold change p-value
padj_shrink: shrinkage of log2 fold change adjusted p-value
se_merge_name: merged SE names
}
\description{
A function to get the category and ranking of SEs in one step
}
\details{
This function is the main function used to get the category and ranking of SEs.
The workflow of this function is "SEfilter -> enhancerFoldchange -> SEfitspline -> SEpattern -> SEcategory"
}
\examples{
# no blacklist with permutation 10 times with b-spline function
se_main_list <- DASE(se_in=pooled_se,e_in=pooled_enhancer,
condition_1="s1_r1_path,s1_r2_path",
condition_2="s2_r1_path,s2_r2_path",
c1_pair="F,F",c2_pair="F,F")

# with blacklist and permutation 10 times with n-spline function
se_main_list <- DASE(se_in=pooled_se,e_in=pooled_enhancer,bl_file= blacklist,
s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path,
spline_fun="ns")

# no blacklist nor permutation 10 times with smooth spline function
se_main_list <- DASE(se_in=pooled_se,e_in=pooled_enhancer,permut=FALSE,
s1_r1_bam=s1_r1_path,s1_r2_bam=s1_r2_path,s2_r1_bam=s2_r1_path,s2_r2_bam=s2_r2_path,
spline_fun="smooth")

}
