% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEfitspline.R
\name{SEfitspline}
\alias{SEfitspline}
\title{Fit enhancer fold change and normal premut}
\usage{
SEfitspline(e_deseq_out_df)
}
\arguments{
\item{e_deseq_out_df}{enhancer output file from enhancerFoldchange}
}
\value{
se_fit_df: enhancer fitted dataframe
spline_plot_df_out: spline fitted line plot dataframe
}
\description{
A function uses bs-spline weighted by basemean to fit the enhancer fold change
within each SE.
}
\details{
This function will fit the enhancer fold change within each SE by using bs-spline
weighted by the max basemean of two samples.
}
\examples{
fit_list <- SEfitspline(enhancerFoldchange_out)

}
