% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEfitspline_bs.R
\name{SEfitspline_bs}
\alias{SEfitspline_bs}
\title{Fit enhancer fold change and normal premut}
\usage{
SEfitspline_bs(e_deseq_out_df, c1_n, c2_n)
}
\arguments{
\item{e_deseq_out_df}{enhancer output file from enhancerFoldchange}

\item{c1_n}{number of replicates (samples) in condition 1.}

\item{c2_n}{number of replicates (samples) in condition 2.}
}
\value{
se_fit_df: enhancer fitted dataframe
spline_plot_df_out: spline fitted line plot dataframe

Each column in se_fit_df represents:

e_merge_name: name of merged CEs.
chr: chromosome of CEs.
start: start posation of CEs.
end: end posation of CEs.
width: width of CEs.
C1_1: coverage of condition 1 replicate 1.
C1_2: coverage of condition 1 replicate 2.
C2_1: coverage of condition 2 replicate 1.
C2_2: coverage of condition 2 replicate 2.
C1_1_norm: normalized coverage of condition 1 replicate 1.
C1_2_norm: normalized coverage of condition 1 replicate 2.
C2_1_norm: normalized coverage of condition 2 replicate 1.
C2_2_norm: normalized coverage of condition 2 replicate 2.
baseMean: normalized coverage mean of all samples.
log2FoldChange: log2 fold change of normalized coverage.
lfcSE: log2 fold change standard error.
stat: statistic of log2 fold change.
pvalue: p-value of log2 fold change.
padj: adjusted p-value of log2 fold change.
baseMean_shrink: shrinkage of normalized coverage mean of all samples based on the methods of \emph{DESeq2}.
log2FoldChange_shrink: shrinkage of log2 fold change..
lfcSE_shrink: shrinkage of log2 fold change standard error.
pvalue_shrink: shrinkage of log2 fold change p-value
padj_shrink: shrinkage of log2 fold change adjusted p-value
se_merge_name: merged SE names
}
\description{
A function uses b-spline weighted by basemean to fit the enhancer fold change
within each SE.
}
\details{
This function will fit the enhancer fold change within each SE by using b-spline
weighted by the max basemean of two samples.
}
\examples{
fit_list <- SEfitspline_bs(enhancerFoldchange_out,c1_n=2,c2_n=2)

}
