% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringent_permut.R
\name{stringent_permut}
\alias{stringent_permut}
\title{Fit enhancer fold change and stringent premut with all enhancers}
\usage{
stringent_permut(
  not_in_se_count_matrix,
  e_deseq_out_df,
  se_df,
  permut = T,
  times = 10
)
}
\arguments{
\item{not_in_se_count_matrix}{enhancer count matrix not in SE range
(header "e_merge_name","S1_r1","S1_r2","S2_r1","S2_r2","se_merge_name","chr","start","end")}

\item{e_deseq_out_df}{enhancer output file from enhancerFoldchange}

\item{se_df}{merged SE metadata from SEfilter}

\item{permut}{if you want permutation (default=TRUE)}

\item{times}{permutation times (default=10)}
}
\value{
A list of 4 datasets: enhancer fitted dataset, permutation dataset, cutoff vector,
and a density plot of permutation and original fitted fold change.
}
\description{
A function uses bs-spline weighted by basemean to fit the enhancer fold change
within each SE.
}
\details{
This function will fit the enhancer fold change within each SE by using bs-spline
weighted by basemean. A stringent permutation of counts of all enhancers is used to decide the
cutoff of significant log2FC (for samples are too similar). Cutoff was calculate by slope equal to range(y)/range(x).
The max value in range(-2,-1.5) and min value in range(1.5,2) are set to be lower and upper cutoff.
}
\examples{
#' # permutation 10 times
fit_list <- stringent_permut(pool_enhancer_df,enhancerFoldchange_out,se_meta,
s1_r1_bam,s1_r2_bam,s2_r1_bam,s2_r2_bam)

# permutation 5 times
fit_list <- stringent_permut(pool_enhancer_df,enhancerFoldchange_out,se_meta,times=5,
s1_r1_bam,s1_r2_bam,s2_r1_bam,s2_r2_bam)

# no permutation
fit_list <- stringent_permut(pool_enhancer_df,enhancerFoldchange_out,se_meta,permut=FALSE,
s1_r1_bam,s1_r2_bam,s2_r1_bam,s2_r2_bam)

}
